+++ src/static/current-state/index.html
@@ -41,6 +41,12 @@
       box-sizing: border-box;
     }
 
+    table {
+      table-layout: auto;
+      border-collapse: separate;
+      border-spacing: 0 5px;
+    }
+
     h3 {
       font-weight: 400;
     }
